Updated Game logic at start of game
The game now creates the bombs after the first click, making sure that the first square clicked on doesn't contain a number. 
Also revamped how coordinates in matrices work, before they were represented by a grid [][], now they are represented by a singluar vector []. 

It is also now possible to write an image to the canvas instead of a number with the new writeSprite function.